refactor : 로또 결과확인 로직 수정
Lottos 내에서 검증할 수 있도록 수정 진행
