+++ src/main/kotlin/camp/nextstep/edu/step/step2/domain/lotto/Lottos.kt
@@ -1,5 +1,6 @@
 package camp.nextstep.edu.step.step2.domain.lotto
 
+import camp.nextstep.edu.step.step2.domain.result.LottoMatch
 import java.util.stream.Collectors
 
 class Lottos(
@@ -8,6 +9,17 @@ class Lottos(
 
     init {
         require(lottos.isNotEmpty()) { "로또는 1개 이상이어야 합니다." }
+    }
+
+    fun checkLottoNumbersByWinningLotto(winningLotto: WinningLotto): List<LottoMatch> {
+        val lottoMatchResults = mutableListOf<LottoMatch>()
+
+        for (lotto in lottos) {
+            val matchNumbers = lotto.countMatch(winningLotto.winningLotto)
+            val matchBonus = lotto.countMatchBonus(winningLotto.bonusNumber)
+            lottoMatchResults.add(LottoMatch.of(matchNumbers, matchBonus))
+        }
+        return lottoMatchResults
     }
 
     fun getLottoElements(): List<List<Int>> {
